+++ platform-sdk/consensus-otter-tests/src/testFixtures/java/org/hiero/otter/fixtures/container/ContainerTestEnvironment.java
@@ -50,7 +50,7 @@ public class ContainerTestEnvironment implements TestEnvironment {
      * Constructor with default values for using random node-ids and default directory for container logs.
      */
     public ContainerTestEnvironment() {
-        this(true, getDefaultOutputDirectory(ENV_NAME));
+        this(true, getDefaultOutputDirectory(ENV_NAME), true);
     }
 
     /**
@@ -59,16 +59,18 @@ public class ContainerTestEnvironment implements TestEnvironment {
      * @param useRandomNodeIds {@code true} if the node IDs should be selected randomly; {@code false} otherwise
      */
     public ContainerTestEnvironment(final boolean useRandomNodeIds) {
-        this(useRandomNodeIds, getDefaultOutputDirectory(ENV_NAME));
+        this(useRandomNodeIds, getDefaultOutputDirectory(ENV_NAME), true);
     }
 
     /**
      * Constructor for the {@link ContainerTestEnvironment} class with custom output directory.
      *
-     * @param useRandomNodeIds {@code true} if the node IDs should be selected randomly; {@code false} otherwise
+     * @param useRandomNodeIds    {@code true} if the node IDs should be selected randomly; {@code false} otherwise
      * @param rootOutputDirectory the root directory where container logs will be written per test
+     * @param proxyEnabled        {@code true} if the toxiproxy should be enabled; {@code false} otherwise
      */
-    public ContainerTestEnvironment(final boolean useRandomNodeIds, @NonNull final Path rootOutputDirectory) {
+    public ContainerTestEnvironment(
+            final boolean useRandomNodeIds, @NonNull final Path rootOutputDirectory, final boolean proxyEnabled) {
         ContainerLogConfigBuilder.configure();
 
         this.rootOutputDirectory = rootOutputDirectory;
@@ -82,7 +84,8 @@ public class ContainerTestEnvironment implements TestEnvironment {
             fail("Failed to prepare directory: " + rootOutputDirectory, ex);
         }
 
-        network = new ContainerNetwork(timeManager, transactionGenerator, rootOutputDirectory, useRandomNodeIds);
+        network = new ContainerNetwork(
+                timeManager, transactionGenerator, rootOutputDirectory, useRandomNodeIds, proxyEnabled);
     }
 
     /**
