+++ platform-sdk/consensus-otter-tests/src/testFixtures/java/org/hiero/otter/fixtures/internal/AbstractNetwork.java
@@ -144,8 +144,6 @@ public abstract class AbstractNetwork implements Network {
 
     private NodeId nextNodeId = NodeId.FIRST_NODE_ID;
 
-    protected boolean proxyDisabled;
-
     protected AbstractNetwork(@NonNull final Random random, final boolean useRandomNodeIds) {
         this.random = requireNonNull(random);
         this.useRandomNodeIds = useRandomNodeIds;
@@ -306,14 +304,6 @@ public abstract class AbstractNetwork implements Network {
 
         nextNodeId = nextNodeId.getOffset(randomAdvance + 1L);
         return nextId;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public void disableProxy() {
-        proxyDisabled = true;
     }
 
     /**
